move scope and add tests
